Simplify architecture
