+++ Cargo.toml
@@ -14,8 +14,6 @@ members = [
   "subduction_ephemeral",
   "subduction_http_longpoll",
   "subduction_iroh",
-  "subduction_keyhive_orchestrator",
-  "subduction_keyhive_policy",
   "subduction_keyhive",
   "subduction_wasm",
   "subduction_websocket"
@@ -50,9 +48,7 @@ subduction_ephemeral = { version = "0.6.2", path = "subduction_ephemeral", defau
 subduction_http_longpoll = { version = "0.9.1", path = "subduction_http_longpoll", default-features = false }
 subduction_iroh = { version = "0.8.1", path = "subduction_iroh" }
 subduction_keyhive = { version = "0.5.1", path = "subduction_keyhive" }
-subduction_keyhive_orchestrator = { version = "0.1.0", path = "subduction_keyhive_orchestrator" }
-subduction_keyhive_policy = { version = "0.9.1", path = "subduction_keyhive_policy" }
-subduction_wasm = { version = "0.16.1", path = "subduction_wasm", default-features = false }
+subduction_wasm = { version = "0.16.0", path = "subduction_wasm", default-features = false }
 subduction_websocket = { version = "0.10.1", path = "subduction_websocket", default-features = false }
 
 # External crates
