+++ subduction_keyhive/src/cache.rs
@@ -4,35 +4,28 @@
 //!   event hashes. Rebuilt on each refresh and atomically swapped.
 //! - Global byte stores: keyed by hash, deduplicated across all
 //!   agents. Events are immutable so these grow monotonically.
-//! - Pending op hashes: locally generated ops not yet on a
-//!   confirmed syncpoint. Snapshotted so the wire path doesn't
-//!   re-lock keyhive on every sync request.
 //!
 //! [`hashes_for_peer_pair`](PeriodicEventCache::hashes_for_peer_pair)
 //! joins the per-agent sets with the global byte stores at request time.
 
-use alloc::{
-    collections::{BTreeMap, BTreeSet, btree_map::Entry},
-    vec::Vec,
-};
+use alloc::collections::{BTreeMap, BTreeSet, btree_map::Entry};
 
 use keyhive_core::{
     listener::membership::MembershipListener,
     store::ciphertext::{CiphertextStore, CiphertextStoreExt},
 };
 use keyhive_crypto::{content::reference::ContentRef, signer::async_signer::AsyncSigner};
-use subduction_keyhive::{
+
+use crate::{
     KeyhivePeerId, KeyhiveProtocol, ProtocolError,
     connection::KeyhiveConnection,
-    message::{AgentHashMap, EventHash},
+    message::{AgentHashMap, CborBytes, EventBytes, EventHash},
     storage::KeyhiveStorage,
 };
 
-pub use subduction_keyhive::message::{CborBytes, EventBytes};
-
 /// Periodic event cache.
 #[derive(Debug, Default)]
-pub struct PeriodicEventCache {
+pub(crate) struct PeriodicEventCache {
     /// Per-agent hash sets.
     ///
     /// For any peer pair `(a, b)`, the relevant event set is
@@ -44,12 +37,6 @@ pub struct PeriodicEventCache {
     ///
     /// Unioned into every peer pair's hash set without intersection.
     public_hashes: BTreeSet<EventHash>,
-
-    /// Locally generated ops not yet on a confirmed syncpoint.
-    ///
-    /// Snapshotted by refresh so the protocol's sync-request path
-    /// doesn't re-lock keyhive on every wire roundtrip.
-    pending_op_hashes: Vec<EventHash>,
 
     /// Hash -> raw event bytes paired with pre-encoded CBOR byte-string
     /// framing. Monotonic growth across refreshes. Events are immutable,
@@ -63,65 +50,38 @@ pub struct PeriodicEventCache {
 
 impl PeriodicEventCache {
     /// Create an empty cache.
-    #[must_use]
-    pub const fn new() -> Self {
+    pub(crate) const fn new() -> Self {
         Self {
             agent_hashes: BTreeMap::new(),
             public_hashes: BTreeSet::new(),
-            pending_op_hashes: Vec::new(),
             event_data: BTreeMap::new(),
             last_total_ops: None,
         }
     }
 
     /// Whether the cache has no agent entries and no public hashes.
-    #[must_use]
-    pub fn is_empty(&self) -> bool {
+    #[cfg(test)]
+    pub(crate) fn is_empty(&self) -> bool {
         self.agent_hashes.is_empty() && self.public_hashes.is_empty()
     }
 
     /// Total number of tracked agents (excluding the public agent,
     /// which lives in its own field).
-    #[must_use]
-    pub fn agent_count(&self) -> usize {
+    #[cfg(test)]
+    pub(crate) fn agent_count(&self) -> usize {
         self.agent_hashes.len()
     }
 
-    /// Look up cached per-agent hashes for a peer, if known.
-    #[must_use]
-    pub fn hashes_for_agent(&self, peer: &KeyhivePeerId) -> Option<&BTreeSet<EventHash>> {
-        self.agent_hashes.get(peer)
-    }
-
     /// Cached hashes accessible to the public agent.
-    #[must_use]
-    pub const fn public_hashes(&self) -> &BTreeSet<EventHash> {
+    #[cfg(test)]
+    pub(crate) const fn public_hashes(&self) -> &BTreeSet<EventHash> {
         &self.public_hashes
     }
 
-    /// Cached pending-op hashes (locally generated, awaiting
-    /// confirmation).
-    #[must_use]
-    pub fn pending_op_hashes(&self) -> &[EventHash] {
-        &self.pending_op_hashes
-    }
-
     /// Last observed keyhive total-ops count.
-    #[must_use]
-    pub const fn last_total_ops(&self) -> Option<u64> {
+    #[cfg(test)]
+    pub(crate) const fn last_total_ops(&self) -> Option<u64> {
         self.last_total_ops
-    }
-
-    /// Look up the raw event bytes for a hash, if cached.
-    #[must_use]
-    pub fn event_bytes_for(&self, hash: &EventHash) -> Option<&EventBytes> {
-        self.event_data.get(hash).map(|(eb, _)| eb)
-    }
-
-    /// Look up the pre-encoded CBOR bytes for a hash, if cached.
-    #[must_use]
-    pub fn cbor_bytes_for(&self, hash: &EventHash) -> Option<&CborBytes> {
-        self.event_data.get(hash).map(|(_, cb)| cb)
     }
 
     /// Public hashes ∪ (A ∩ B).
@@ -129,8 +89,11 @@ impl PeriodicEventCache {
     /// Every event the public agent can see is unconditionally included,
     /// plus events reachable to *both* `a` and `b`. Returns a map keyed
     /// by hash with `(EventBytes, CborBytes)` cloned from the byte stores.
-    #[must_use]
-    pub fn hashes_for_peer_pair(&self, a: &KeyhivePeerId, b: &KeyhivePeerId) -> AgentHashMap {
+    pub(crate) fn hashes_for_peer_pair(
+        &self,
+        a: &KeyhivePeerId,
+        b: &KeyhivePeerId,
+    ) -> AgentHashMap {
         let mut out = AgentHashMap::new();
         for h in &self.public_hashes {
             self.insert_pair_entry(&mut out, *h);
@@ -162,7 +125,7 @@ impl PeriodicEventCache {
     /// # Errors
     ///
     /// Returns [`ProtocolError`] if event serialization fails.
-    pub async fn refresh<Signer, T, P, C, L, R, Conn, Store, K>(
+    pub(crate) async fn refresh<Signer, T, P, C, L, R, Conn, Store, K>(
         &mut self,
         protocol: &KeyhiveProtocol<Signer, T, P, C, L, R, Conn, Store, K>,
     ) -> Result<bool, ProtocolError<Conn::SendError>>
@@ -198,11 +161,8 @@ impl PeriodicEventCache {
             KeyhivePeerId::from_identifier(&keyhive_core::principal::public::Public.id());
         let new_public = new_agent_hashes.remove(&public_peer).unwrap_or_default();
 
-        let new_pending = protocol.get_pending_hashes().await?;
-
         self.agent_hashes = new_agent_hashes;
         self.public_hashes = new_public;
-        self.pending_op_hashes = new_pending;
         self.last_total_ops = Some(total);
         Ok(true)
     }
