--- subduction_keyhive_orchestrator/Cargo.toml
@@ -1,43 +1,0 @@
-[package]
-name = "subduction_keyhive_orchestrator"
-version = "0.1.0"
-description = "Orchestration layer (syncpoints, cache, rate limiting, storage policy) above the keyhive sync protocol"
-readme = "./README.md"
-
-categories = ["authentication", "cryptography"]
-keywords = ["sync", "keyhive", "orchestration", "subduction"]
-
-authors.workspace = true
-license.workspace = true
-edition.workspace = true
-repository.workspace = true
-rust-version.workspace = true
-
-[dependencies]
-async-lock = { workspace = true }
-future_form = { workspace = true }
-futures = { workspace = true }
-keyhive_core = { workspace = true }
-keyhive_crypto = { workspace = true }
-rand = { workspace = true }
-sedimentree_core = { workspace = true, default-features = false }
-serde = { workspace = true }
-subduction_core = { workspace = true, default-features = false }
-subduction_crypto = { workspace = true }
-subduction_keyhive = { workspace = true, features = ["serde"] }
-thiserror = { workspace = true }
-tracing = { workspace = true }
-
-[dev-dependencies]
-async-channel = { workspace = true }
-ciborium = { workspace = true }
-subduction_keyhive = { workspace = true, features = ["serde", "test-utils"] }
-testresult = { workspace = true }
-tokio = { workspace = true, features = ["macros", "rt"] }
-
-[features]
-default = []
-std = ["subduction_core/std"]
-
-[lints]
-workspace = true
