--- subduction_keyhive_policy/Cargo.toml
@@ -1,43 +1,0 @@
-[package]
-name = "subduction_keyhive_policy"
-version = "0.9.1"
-description = "Keyhive-based authorization policy for Subduction connections"
-readme = "./README.md"
-
-
-categories = ["authentication", "cryptography"]
-keywords = ["authorization", "policy", "sync"]
-
-authors.workspace = true
-license.workspace = true
-edition.workspace = true
-repository.workspace = true
-rust-version.workspace = true
-
-[dependencies]
-async-channel = { workspace = true }
-ed25519-dalek = { workspace = true }
-futures = { workspace = true }
-future_form = { workspace = true }
-keyhive_core = { workspace = true }
-keyhive_crypto = { workspace = true }
-rand = { workspace = true }
-sedimentree_core = { workspace = true, default-features = false }
-serde = { workspace = true }
-subduction_core = { workspace = true, default-features = false }
-subduction_crypto = { workspace = true }
-subduction_keyhive = { workspace = true, features = ["serde"] }
-thiserror = { workspace = true }
-tracing = { workspace = true }
-
-[dev-dependencies]
-subduction_core = { workspace = true, features = ["test_utils"] }
-testresult = { workspace = true }
-tokio = { workspace = true, features = ["macros", "rt", "time"] }
-
-[features]
-default = []
-std = ["subduction_core/std"]
-
-[lints]
-workspace = true
