--- subduction_keyhive_policy/README.md
@@ -1,11 +1,0 @@
-# subduction_keyhive_policy
-
-Keyhive-based authorization policy for Subduction connections.
-
-## Overview
-
-This crate provides a `Policy` implementation for Subduction that uses Keyhive for connection authorization, enabling fine-grained access control over sync operations.
-
-## License
-
-Licensed under either of Apache License, Version 2.0 or MIT license at your option.
